# Replace cnf structhash with custom one