This version has few improvements:

- the most important thing - it only dumps struct fields that **have** hash:name tag (it allows us to skip all XXX_ generated fields by protobuf)
- cover edge cases for dump maps/interface
- 100% test coverage 
- better performance

```
➜  structhash git:(feature/new-structhash) ✗ benchcmp old.out new.out
benchmark                 old ns/op     new ns/op     delta
BenchmarkSimpleJSON-6     852           894           +4.93%
BenchmarkSimpleDump-6     3806          1145          -69.92%
BenchmarkFullJSON-6       5888          6110          +3.77%
BenchmarkFullDump-6       20447         1168          -94.29%
BenchmarkTagsJSON-6       572           594           +3.85%
BenchmarkTagsDump-6       3433          4489          +30.76%

benchmark                 old allocs     new allocs     delta
BenchmarkSimpleJSON-6     1              1              +0.00%
BenchmarkSimpleDump-6     22             12             -45.45%
BenchmarkFullJSON-6       9              9              +0.00%
BenchmarkFullDump-6       130            12             -90.77%
BenchmarkTagsJSON-6       1              1              +0.00%
BenchmarkTagsDump-6       21             34             +61.90%

benchmark                 old bytes     new bytes     delta
BenchmarkSimpleJSON-6     96            96            +0.00%
BenchmarkSimpleDump-6     1056          216           -79.55%
BenchmarkFullJSON-6       960           960           +0.00%
BenchmarkFullDump-6       6336          216           -96.59%
BenchmarkTagsJSON-6       48            48            +0.00%
BenchmarkTagsDump-6       680           960           +41.18%
```